Move OS-specific APIs to a separate header
