+++ include/fmt/core.h
@@ -1503,8 +1503,7 @@ FMT_API void vprint(string_view format_str, format_args args);
 /**
   \rst
   Prints formatted data to the file *f*. For wide format strings,
-  *f* should be in wide-oriented mode set via ``fwide(f, 1)`` or
-  ``_setmode(_fileno(f), _O_U8TEXT)`` on Windows.
+  *f* should be in wide-oriented mode set via ``fwide(f, 1)``.
 
   **Example**::
